--- Testing/Exec/ReactEval/compareCantera/canteraSim/mechanism/prepareXml_dodecane_lu_qss.sh
@@ -1,3 +1,0 @@
-ck2cti --input=dodecane_lu_qss.inp --thermo=dodecane_lutherm.dat --transport=dodecane_lutran.dat --permissive
-python -m cantera.cti2yaml dodecane_lu_qss.cti
-
