--- Testing/Exec/ReactEval/compareCantera/canteraSim/mechanism/prepareXml_drm19.sh
@@ -1,3 +1,0 @@
-ck2cti --input=drm19.inp --thermo=drm19therm.dat --transport=drm19tran.dat --permissive
-python -m cantera.cti2yaml drm19.cti
-
